+++ WebQuanLySinhVien/Models/ViewModels/ThongKe/SL_SinhVienNganh.cs
@@ -7,7 +7,7 @@
         public string? SelectedID { get; set; }
         public string? SelectedSVID { get; set; }
         public string? IdNganh { get; set; }
-        public int[] DanhSachDiem { get; set; }
+        public decimal[] DanhSachDiem { get; set; }
         public IList<string>? DanhSachIdHocPhan { get; set; }
         public List<SoLuongSinhVien> DanhSachSoLuongSinhVien { get; set; }
     }
